docs: Update the Node.js get started guide
This commit updates the Node.js get started guide:
- assumes the user has an existing application
and provides optional instructions to create
an example application
- adds sample code for CommonJS applications
