+++ docs/current/sdk/nodejs/snippets/get-started/step2/build.mts
