+++ docs/current/sdk/nodejs/snippets/get-started/step3/build.mts
