+++ docs/current/sdk/nodejs/snippets/get-started/step4/build.js
@@ -1,29 +1,43 @@
-import { connect } from "@dagger.io/dagger"
+;(async function () {
+  // initialize Dagger client
+  let connect = (await import("@dagger.io/dagger")).connect
 
-// initialize Dagger client
-connect(async (client) => {
-  // highlight-start
-  // get reference to the local project
-  const source = client.host().directory(".", { exclude: ["node_modules/"] })
+  connect(async (client) => {
+    // highlight-start
+    // Set Node versions against which to test and build
+    const nodeVersions = ["12", "14", "16"]
+    // highlight-end
 
-  // get Node image
-  const node = client.container().from("node:16")
+    // get reference to the local project
+    const source = client.host().directory(".", { exclude: ["node_modules/"] })
 
-  // mount cloned repository into Node image
-  const runner = client
-    .container({ id: node })
-    .withMountedDirectory("/src", source)
-    .withWorkdir("/src")
-    .withExec(["npm", "install"])
+    // highlight-start
+    // for each Node version
+    for (const nodeVersion of nodeVersions) {
+      // get Node image
+      const node = client.container().from(`node:${nodeVersion}`)
+      // highlight-end
 
-  // run tests
-  await runner.withExec(["npm", "test", "--", "--watchAll=false"]).exitCode()
+      // mount cloned repository into Node image
+      const runner = client
+        .container({ id: node })
+        .withMountedDirectory("/src", source)
+        .withWorkdir("/src")
+        .withExec(["npm", "install"])
 
-  // build application
-  // write the build output to the host
-  await runner
-    .withExec(["npm", "run", "build"])
-    .directory("build/")
-    .export("./build")
-  // highlight-end
+      // run tests
+      await runner
+        .withExec(["npm", "test", "--", "--watchAll=false"])
+        .exitCode()
+
+      // highlight-start
+      // build application using specified Node version
+      // write the build output to the host
+      await runner
+        .withExec(["npm", "run", "build"])
+        .directory("build/")
+        .export(`./build-node-${nodeVersion}`)
+    }
+    // highlight-end
+  })
 })
