+++ docs/current/sdk/nodejs/snippets/get-started/step4/build.mjs
